moving the xga information to the misc file
